app/requirements.txt
requests==2.32.5
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytest==8.4.2
requests-mock==1.12.1
pytest-postgresql==7.0.2
